+++ Changelog.md
@@ -1,10 +1,20 @@
 Changelog
 =========
 
-Version 0.6.0
+Version 0.7.0
 -------------
 * added direction rendering to relations in class and state machine views
 * enhanced example models
+
+
+Version 0.6.0
+-------------
+* refactored exports, distinguish between
+  * exports of model data (to JSON, structurizr)
+  * rendering of views (e.g. to PlantUML)
+* changed command line options to reflect the refactoring
+* added render-format 'all' to generate all formats in one go
+* updated usage and design documentation and diagrams
 
 
 Version 0.5.0